Fix critical memory safety, security, and functional issues
Critical fixes:
- Fix use-after-free in screen capture by using Arc instead of Box
- Fix memory leak in video decoder by removing std::mem::forget
- Replace certificate skip with Trust-On-First-Use (TOFU) verification

Important fixes:
- Implement Bonjour service discovery using DNS-SD APIs
- Implement H.264/H.265 NAL unit parsing (AVCC and Annex B formats)
- Add proper VideoFrameHeader with frame fragmentation for large frames
- Complete Metal display with CAMetalLayer integration
- Add graceful shutdown signal handling (SIGTERM/SIGINT)

Minor fixes:
- Fix audio sample count to use actual channel count
- Apply clippy fixes across crates

+++ bridge-input/src/capture.rs
@@ -395,13 +395,13 @@ extern "C" fn event_tap_callback(
 
         CGEventType::ScrollWheel => {
             let delta_y = unsafe {
-                CGEventGetDoubleValueField(event, K_CG_SCROLL_WHEEL_EVENT_DELTA_AXIS_1 as u32)
+                CGEventGetDoubleValueField(event, K_CG_SCROLL_WHEEL_EVENT_DELTA_AXIS_1)
             };
             let delta_x = unsafe {
-                CGEventGetDoubleValueField(event, K_CG_SCROLL_WHEEL_EVENT_DELTA_AXIS_2 as u32)
+                CGEventGetDoubleValueField(event, K_CG_SCROLL_WHEEL_EVENT_DELTA_AXIS_2)
             };
             let is_continuous = unsafe {
-                CGEventGetIntegerValueField(event, K_CG_SCROLL_WHEEL_EVENT_IS_CONTINUOUS as u32) != 0
+                CGEventGetIntegerValueField(event, K_CG_SCROLL_WHEEL_EVENT_IS_CONTINUOUS) != 0
             };
 
             let input_event = InputEvent::MouseScroll(MouseScrollEvent {
